refactor : mvc 패턴 수정
