+++ src/main/kotlin/racing/RacingApplication.kt
@@ -1,7 +1,11 @@
 package racing
 
+import racing.controller.RacingController
+import racing.view.InputView
+import racing.view.ResultView
+
 class RacingApplication
 
 fun main() {
-    Racing().racingGame()
+    RacingController(InputView, ResultView).racingGame()
 }
